+++ src/__tests__/simple/enum.test.ts
@@ -1,5 +1,5 @@
 /* istanbul ignore file */
-import enums from '../enum';
+import enums from '../../enum';
 
 describe('enum.js', () => {
   const supportedComparisonOperator = [
