+++ src/__tests__/simple/examples/firstLambda/event.json
