+++ src/__tests__/simple/examples/firstLambda/index.js
